Handling errors;
Changing vault import to generated by typechain;
creating a unit tests file;
changing name of integration tests file;
