+++ balancer-js/src/balancerErrors.ts
@@ -31,6 +31,7 @@ export enum BalancerErrorCode {
   GAUGES_REWARD_TOKEN_EMPTY = 'GAUGES_REWARD_TOKEN_EMPTY',
   REWARD_TOKEN_ZERO = 'REWARD_TOKEN_ZERO',
   FEE_PROVIDER_NOT_PROVIDED = 'FEE_PROVIDER_NOT_PROVIDED',
+  MIN_SWAP_FEE_PERCENTAGE = 'MIN_SWAP_FEE_PERCENTAGE',
 }
 
 export class BalancerError extends Error {
