+++ balancer-js/src/modules/pools/factory/weighted/weighted.factory.spec.ts
@@ -6,7 +6,7 @@ import { Network, PoolType } from '@/types';
 import { ADDRESSES } from '@/test/lib/constants';
 import { BalancerSDK } from '@/modules/sdk.module';
 import { Interface, LogDescription } from '@ethersproject/abi';
-import { forkSetup } from '@/test/lib/utils';
+import { findEventInReceiptLogs, forkSetup } from '@/test/lib/utils';
 import dotenv from 'dotenv';
 import { isSameAddress } from '@/lib/utils';
 import { Vault__factory } from '@/contracts/factories/Vault__factory';
@@ -20,7 +20,7 @@ dotenv.config();
 
 const network = Network.MAINNET;
 const rpcUrl = 'http://127.0.0.1:8545';
-const alchemyRpcUrl = `${process.env.ALCHEMY_URL}`;
+const alchemyRpcUrl = `${ process.env.ALCHEMY_URL }`;
 const blockNumber = 16320000;
 
 const name = 'My-Test-Pool-Name';
@@ -31,11 +31,11 @@ const addresses = ADDRESSES[network];
 const USDC_address = addresses.USDC.address;
 const USDT_address = addresses.USDT.address;
 
-const factoryAddress = `${BALANCER_NETWORK_CONFIG[network].addresses.contracts.weightedPoolFactory}`;
+const factoryAddress = `${ BALANCER_NETWORK_CONFIG[network].addresses.contracts.weightedPoolFactory }`;
 const owner = '0xf39Fd6e51aad88F6F4ce6aB8827279cffFb92266';
 const tokenAddresses = [USDC_address, USDT_address];
 const swapFee = '0.01';
-const weights = [`${0.2e18}`, `${0.8e18}`];
+const weights = [`${ 0.2e18 }`, `${ 0.8e18 }`];
 const slots = [addresses.USDC.slot, addresses.USDT.slot];
 const balances = [
   parseFixed('100000', 6).toString(),
@@ -90,14 +90,9 @@ describe('creating weighted pool', () => {
         WeightedPoolFactory__factory.abi
       );
 
-      const poolCreationEvent: LogDescription | null | undefined = receipt.logs
-        .filter((log: Log) => {
-          return isSameAddress(log.address, factoryAddress);
-        })
-        .map((log) => {
-          return weightedPoolFactoryInterface.parseLog(log);
-        })
-        .find((parsedLog) => parsedLog?.name === 'PoolCreated');
+      const poolCreationEvent: LogDescription = findEventInReceiptLogs({
+        receipt:
+      });
       if (poolCreationEvent) {
         poolAddress = poolCreationEvent.args.pool;
       }
